requirements.txt
black==25.1.0
blinker==1.9.0
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Mail==0.10.0
folium==0.19.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mypy_extensions==1.1.0
numpy==2.2.3
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
xyzservices==2025.1.0
